requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
soupsieve==2.6
yarl==1.18.0
textblob
python-dotenv==1.0.0
tldextract==5.0.1
lxml==4.9.3
requests==2.28.2
nltk==3.8.1
flask==2.2.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
celery==5.2.7
supabase==1.0.3
postgrest-py==0.10.8
gotrue==1.1.0
flask-limiter==3.3.0
email-validator==2.1.0
flask-cors==3.0.10
pydantic>=2.0.0
psycopg2-binary==2.9.6
redis==4.5.4
dataforseo-client==1.0.0
requests==2.28.2
